% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_niche.R
\name{apply_niche}
\alias{apply_niche}
\title{apply niche preference}
\usage{
apply_niche(x, niche_def, gc)
}
\arguments{
\item{x}{events, e.g. times/ages of fossil occurrences or their stratigraphic position.}

\item{niche_def}{function, specifying the niche along a gradient. Should return 0 when taxon is outside of niche, and 1 when fully inside niche. Values between 0 and 1 are interpreted as probabilities.}

\item{gc}{function, stands for "gradient change". Specifies how the gradient changes, e.g. with time}
}
\description{
Models niche preferences by removing events (fossil occurrences) when they are outside of a preferred niche using the function \code{thin}.
Combines the functions \code{niche_def} and \code{gc} ("gradient change") to determine how the taxons' preferred environment changes with time. This is done by composing \code{niche_def} and \code{gc}. The result is then used as a thinning on the events \code{x}.
}
\examples{
\dontrun{
## setup
# using water depth as gradient
 t = scenarioA$t_myr
 wd = scenarioA$wd_m[,"8km"]
 gc = approxfun(t, wd)
 plot(t, gc(t), type = "l", xlab = "Time", ylab = "water depth [m]",
  main = "gradient change with time")
 # define niche
 # preferred wd 10 m, tolerant to intermediate wd changes (standard deviation 10 m), non-terrestrial
 niche_def = cnd(mean = 10, sd = 10, inc  = 40, cut_neg = TRUE)
 plot(seq(-1, 50, by = 0.5), niche_def(seq(-1, 50, by = 0.5)), type = "l",
 xlab = "water depth", ylab = "preference", main = "Niche def")
 # niche pref with time
 plot(t, niche_def(gc(t)), type = "l", xlab = "time", ylab = "preference", main = "pref with time")

 ## simulate fossil occurrences
 foss_occ = p3(rate = 100, from = 0, to = max(t))
 # foss occ without niche pref
 hist(foss_occ, xlab = "time")
 foss_occ_niche = apply_niche(foss_occ, niche_def, gc)
 # fossil occurrences with niche preference
 hist(foss_occ_niche, xlab = "time")

 # see also
 vignette("event_data")
 # for a longer example

}

}
\seealso{
\code{\link[=apply_taphonomy]{apply_taphonomy()}} to model taphonomic effects based on the same principle, \code{\link[=thin]{thin()}} for the underlying mathematical procedure.
}
