% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_char_mat.R
\name{as_char_mat}
\alias{as_char_mat}
\title{construct character matrix}
\usage{
as_char_mat(mat, pos = NULL)
}
\arguments{
\item{mat}{matrix containing the characters, where rows correspond to
taxa and columns correspond to the characters. Must have column names with taxon names}

\item{pos}{named vector, with names matching the taxon names of the matrix. Provides the position (time/height) of the taxa in the character matrix}
}
\value{
a object of S3 class \code{char_mat}, describing a character matrix with associated positions (times/heights) where the taxa were found
}
\description{
constructor for character matrices for stratigraphic transformations & taphonomic effects
}
