% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stratlist.R
\name{as_stratlist}
\alias{as_stratlist}
\title{transform list into strat lisgt}
\usage{
as_stratlist(x)
}
\arguments{
\item{x}{a list}
}
\value{
a list of class "stratlist"
}
\description{
transform list into strat lisgt
}
