% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_timelist.R
\name{as_timelist}
\alias{as_timelist}
\title{convert list to time list for plotting}
\usage{
as_timelist(x)
}
\arguments{
\item{x}{a list}
}
\value{
an object of class time list
}
\description{
convert list to time list for plotting
}
