% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd.R
\name{cnd}
\alias{cnd}
\title{capped normal distribution pdf}
\usage{
cnd(mean, sd, inc = 1, cut_neg = TRUE)
}
\arguments{
\item{mean}{mean of normal distribution}

\item{sd}{standard deviation}

\item{inc}{scalar, factor by which the pdf in multiplied}

\item{cut_neg}{logical. should negative values be set to 0?}
}
\value{
a function
}
\description{
returns a function that defines a niche based on a capped normal distribution, i.e. a pdf of a normal distribution where all values above 1 are capped. Mathematically, this is f(x) = min( inc * pdf(x), 1)
}
