% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_filter.R
\name{gap_filter}
\alias{gap_filter}
\title{remove data associated with gaps}
\usage{
gap_filter(x, adm, ...)
}
\arguments{
\item{x}{data, e.g., a vector of times or a fossils object}

\item{adm}{an age-depth model}

\item{...}{parameters passed to \code{is_destructive}}
}
\value{
for a numeric vector, a vector with entries that coincide with gaps removed. For a fossils object, a smaller fossils object with fossils that coincide with gaps removed
}
\description{
Removes temporal data that coincides with a gap in the stratigraphic record
Specifically, this is a transformation from the time domain unto itself
}
\examples{
#create age-depth model
adm = admtools::tp_to_adm(t = scenarioA$t_myr, h = scenarioA$h_m[,"2km"])
#simulate fossil occurrences
occ = p3(rate = 100, from = 0, to = 2)
breaks = seq(0, 2, length.out = 30)
hist(occ, main = "baseline", breaks = breaks)
hist(gap_filter(occ, adm), main = "Fossil in gaps removed", breaks = breaks)

}
\seealso{
\itemize{
\item \code{\link[=strat_filter]{strat_filter()}}
\item \code{\link[=apply_taphonomy]{apply_taphonomy()}}
}
}
