% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_char_mat.R
\name{is_char_mat}
\alias{is_char_mat}
\title{validity of character matrices}
\usage{
is_char_mat(x, ...)
}
\arguments{
\item{x}{object to test}

\item{...}{other parameters. Currently unused}
}
\value{
TRUE if \code{x} is a valid character matrix object, throws an informative error message otherwise
}
\description{
tests if \code{x} is a valid character matrix (\code{char_mat} object)
}
