% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3.R
\name{p3}
\alias{p3}
\title{simulate fossil occurrences}
\usage{
p3(rate, from, to, n = NULL)
}
\arguments{
\item{rate}{rate of fossil occurrences (avg. no of fossils per unit)}

\item{from}{lowest possible fossil occurrence}

\item{to}{highest possible fossil occurrence}

\item{n}{integer of NULL (default). Number of fossil occurrences to return. If null, the number is random and determined by the rate parameter}
}
\description{
Simulates fossil occurrences in the interval from, to based on a Poisson point process with rate \code{rate}. If the parameter \code{n} is used, the number of fossils is conditioned to be \code{n}
}
