% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p3_var_rate.R
\name{p3_var_rate}
\alias{p3_var_rate}
\title{simulate variable rate Poisson point process}
\usage{
p3_var_rate(x, y = NULL, from = 0, to = 1, fmax = 1, n = NULL)
}
\arguments{
\item{x}{numeric vector or function. If x is a function, it is used to specify the variable rate. If x is a vector, x and y together specify the variable rate using linear interpolation}

\item{y}{numeric vector. If specified, determines the variable rate. This is done by using linear interpolation between the values of y. Here x specifies the ordinate and y the abscissa}

\item{from}{lower limit on the event}

\item{to}{upper limit on the event}

\item{fmax}{maximum rate in the ovserved interval. Required for rejection sampling, If this value is too low, results will be biased}

\item{n}{NULL or an integer. Number of events drawn. If NULL, the number of events is determined by the rate (specified by x and y). If an integer is passed, n events are returned.}
}
\description{
simulate variable rate Poisson point process
}
