% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.stratlist.R
\name{plot.stratlist}
\alias{plot.stratlist}
\title{plot strat list}
\usage{
\method{plot}{stratlist}(x, orientation = "du", ...)
}
\arguments{
\item{x}{stratlist object}

\item{orientation}{charachter, either "du" (down-up) or "lr" (left-right). Orientation of plotting}
}
\description{
plot strat list
}
