% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_remove.R
\name{prob_remove}
\alias{prob_remove}
\title{probabilistic removal of elements}
\usage{
prob_remove(x, prob)
}
\arguments{
\item{x}{vector}

\item{prob}{number between 0 and 1, probability to preserve elements}
}
\value{
a vector of the same type as x
}
\description{
probabilistic removal of elements from x. For each element, the probability to be preserved is independent and specified by prob
}
\examples{
x = prob_remove(1:10, 0.5)
x
x = prob_remove(1:10, 0.5)
x
}
\seealso{
\itemize{
\item \code{\link[=apply_niche]{apply_niche()}} and \code{\link[=apply_taphonomy]{apply_taphonomy()}} for functions that use this function for transformation of \code{pre_paleoTS} objects
}
}
