% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_to_paleoTS.R
\name{reduce_to_paleoTS}
\alias{reduce_to_paleoTS}
\title{reduce pre-paleoTS format to paleoTS}
\usage{
reduce_to_paleoTS(x, ...)
}
\arguments{
\item{x}{a \code{pre_paleoTS} object}

\item{...}{other options. currently unused}
}
\value{
a \code{paleoTS} object
}
\description{
paleoTS is a format for paleontological time series. It is a summary format where interpopulation variance is provided as a parameter. As a result, taphonomic and ecological effects that act on individual specimens can not be modeled for paleoTS objects. To resolve this, the pre_paleoTS format tracks each specimen individually. This function reduces the pre-paleoTS format into standard paleoTS object, which can be used by the paleoTS package.
}
\examples{
x = stasis_sl(t = 0:5)     # create pre_paleoTS object representing stasis on specimen level
y = reduce_to_paleoTS(x)   # reduce to standard paleoTS format
# now analyses using the paleoTS package can be applied to y

}
\seealso{
\code{\link[=stasis_sl]{stasis_sl()}} to simulate stasis on specimen level (sl), returning an object of call \code{pre_paleoTS}
}
