% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rej_samp.R
\name{rej_samp}
\alias{rej_samp}
\title{rejection sampling}
\usage{
rej_samp(f, x_min, x_max, n = 1L, f_max = 1)
}
\arguments{
\item{f}{function. (pseudo) pdf from which the sample is drawn}

\item{x_min}{scalar. lower limit of the examined interval}

\item{x_max}{scalar. upper limit of the examined interval}

\item{n}{integer. number of samples drawn}

\item{f_max}{maximum value of \code{f} in the interval from \code{x_min} to \code{x_max}. If f attains values larger than \code{f_max} a warning is throw, \code{f_max} is adjusted, and sampling is started again}
}
\value{
numeric vector, samples drawn from the (pseudo) pdf
}
\description{
Rejection sampling from the (pseudo) pdf \code{f} in the interval between \code{x_min} and \code{x_max}. Returns \code{n} samples. Note that values of \code{f} below 0 are capped to zero
}
\examples{

f = sin
x = rej_samp(f, 0, 3*pi, n = 100)
hist(x) # note that no samples are drawn where sin is negative

}
\seealso{
\code{\link[=p3_var_rate]{p3_var_rate()}} for the derived variable rate Poisson point process implementation.
}
