% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioA.R
\docType{data}
\name{scenarioA}
\alias{scenarioA}
\title{scenario A from Hohmann et al 2024}
\format{
A list with 6 elements:
\itemize{
\item \code{t_myr} : numeric vector. timesteps of the simulation in Myr
\item \code{sl_m} : numeric vector. eustatic sea level in m
\item \code{dist_from_shore} : character vector. Distance from shore in km of locations at which the observations were made
\item \code{h_m} : matrix of size length(t_myr) x length(dist_from_shore). Accumulated sediment height in m at examined locations
\item \code{wd_m}: matrix of size length(t_myr) x length(dist_from_shore). Water depth in m at examined locations
\item \code{strat_col}: list with length(dist_from shore) elements. Represents a stratigraphic column. Each element is a list with two elements
\item \code{bed_thickness_m}: numeric vector. Bed thickness in m
\item \code{facies_code} : integer vector. facies code of the bed
}
}
\usage{
scenarioA
}
\description{
scenario A from Hohmann et al 2024
}
\keyword{datasets}
