% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stasis_sl.R
\name{stasis_sl}
\alias{stasis_sl}
\title{simulate phenotypic stasis (pre-paleoTS format)}
\usage{
stasis_sl(t, mean = 0, sd = 1, interpop_var = 1, n_per_sample = 10)
}
\arguments{
\item{t}{times at which the traits are determined}

\item{mean}{mean trait value}

\item{sd}{stictly positive number, standard deviation of traits around the mean}

\item{interpop_var}{interpopulation variance, determines how much specimens from the same population vary}

\item{n_per_sample}{integer, number of specimens sampled per population/sampling locality}
}
\value{
description
}
\description{
simulated stasis as independent, normally distributed random variables with mean \code{mean} and standard deviation \code{sd}, draws \code{n_per_sample} samples from each sampling location (population) that have specified variance \code{interpop_var}
}
\seealso{
\code{\link[=stasis]{stasis()}} for the version that simulated stasis of mean trait values.
}
