% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin.R
\name{thin}
\alias{thin}
\title{thin a series of events (e.g. fossil occurrences)}
\usage{
thin(x, thin)
}
\arguments{
\item{x}{numeric vectors with events (e.g. locations, height, times)}

\item{thin}{a function used for thinning}
}
\description{
Thins a vector of events using the function thin, meaning the probability that the ith event in x is preserved is given by \emph{thin (x(i))}. Values of
\code{thin} below 0 and above 1 are ignored
}
