% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_niche_pref.R
\name{apply_niche_pref}
\alias{apply_niche_pref}
\title{apply niche preference to events}
\usage{
apply_niche_pref(x, niche, temp_change)
}
\arguments{
\item{x}{events}

\item{niche}{function, defining a niche along a gradient}

\item{temp_change}{function, change in the niche with time}
}
\description{
apply niche preference to events
}
