% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ornstein_uhlenbeck.R
\name{ornstein_uhlenbeck}
\alias{ornstein_uhlenbeck}
\title{simulate ornstein-uhlenbeck (OU) process}
\usage{
ornstein_uhlenbeck(t, mu = 0, theta = 1, sigma = 1, y0 = 0)
}
\arguments{
\item{t}{times at which the process is simulated}

\item{mu}{numeric, long term mean}

\item{theta}{numeric, mean reversion speed}

\item{sigma}{positive scalar, strength of randomness}

\item{y0}{scalar. initial value (value of process at the first entry of t)}
}
\value{
a list with two elements: \code{t} and \code{y}. \code{t} is a duplicate of the input \code{t}, \code{y} are the values of the OU process at these times. Outputs are of class \code{timelist} and can thus be plotted directly using \code{plot}, see \code{?plot.timelist}
}
\description{
simulates ornstein-uhlenbeck using the Euler-Maruyame method. The process is simulated on a scale of 0.25 * min(diff(t)) and then interpolated to the values of \code{t}.
}
