% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk.R
\name{random_walk}
\alias{random_walk}
\title{continuous time random walk}
\usage{
random_walk(t, sigma = 1, mu = 0, y0 = 0)
}
\arguments{
\item{t}{numeric vector with strictly increasing elements. Times at which the random walk is evaluated}

\item{sigma}{variance parameter}

\item{mu}{directionality parameter}

\item{y0}{starting value, i.e. value of the random walk at the first entry of \code{t}}
}
\value{
a list with elements \code{t} and \code{y}. \code{t} is a duplicate of the input parameter and is the times at which the random walk is evaluated. \code{y} are the values of the random walk at said times
}
\description{
simulates a random walk as a Brownian drift
}
