% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_walk_sl.R
\name{random_walk_sl}
\alias{random_walk_sl}
\title{simulate random walk (pre-paleoTS format)}
\usage{
random_walk_sl(
  t,
  sigma = 1,
  mu = 0,
  y0 = 0,
  interpop_var = 1,
  n_per_sample = 10
)
}
\arguments{
\item{t}{numeric vector with strictly increasing elements, can be heterodistant. Times at which the random walk is evaluated}

\item{sigma}{positive scalar, variance parameter}

\item{mu}{scalar, directionality parameter}

\item{y0}{scalar, starting value (value of the random walk at the first entry of \code{t})}

\item{interpop_var}{interpopulation variance, determines how many specimens from the same population vary}

\item{n_per_sample}{integer, number of specimens sampled per population/sampling locality}
}
\value{
an object of S3 class \code{pre_paleoTS}, inherits from \code{timelist} and \code{list}. The list has two elements: \code{t}, containing a vector of times of sampling, and \code{vals}, a list of trait values of the same length as \code{t}, with element containing trait values of individual specimens.  This object can be transformed using \code{apply_taphonomy}, \code{apply_niche} or \code{time_to_strat}, and then reduced to a \code{paleoTS} object using \code{reduce_to_paleoTS}. This can then be used to test for different modes of evolution.
}
\description{
Simulates a (continuous time) random walk as a Brownian drift on specimen level
}
\seealso{
\code{\link[=random_walk]{random_walk()}} for the equivalent function to simulate mean trait values
}
