% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stasis.R
\name{stasis}
\alias{stasis}
\title{simulate phenotypic stasis}
\usage{
stasis(t, mean = 0, sd = 1)
}
\arguments{
\item{t}{times at which the phenotype is determined}

\item{mean}{mean value}

\item{sd}{standard deviation}
}
\description{
simulates stasis as independent, normally distributed random variables with mean \code{mean} and standard deviatin \code{sd}
}
\examples{
\dontrun{
 library("admtools") # required for plotting of results
 t = seq(0, 1, by = 0.01)
 l = stasis(t)
 plot(l, type = "l") # plot lineage
 l2 = stasis(t, mean = 0.5, sd = 0.3) # simulate second lineage
 lines(l2$t, l2$y, col = "red") # plot second lineage
}

}
