% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratPal-package.R
\docType{package}
\name{StratPal-package}
\alias{StratPal}
\alias{StratPal-package}
\title{StratPal: Stratigraphic Paleobiology Modeling Pipelines}
\description{
The fossil record is a joint expression of ecological, taphonomic, evolutionary, and stratigraphic processes (Holland and Patzkowsky, 2012, ISBN:978-0226649382). This package allowing to simulate biological processes in the time domain (e.g., trait evolution, fossil abundance, phylogenetic trees), and examine how their expression in the rock record (stratigraphic domain) is influenced based on age-depth models, ecological niche models, and taphonomic effects. Functions simulating common processes used in modeling trait evolution, biostratigraphy or event type data such as first/last occurrences are provided and can be used standalone or as part of a pipeline. The package comes with example data sets and tutorials in several vignettes, which can be used as a template to set up one's own simulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mindthegap-erc.github.io/StratPal/} 
  \item \url{https://github.com/MindTheGap-ERC/StratPal}
  \item Report bugs at \url{https://github.com/MindTheGap-ERC/StratPal/issues}
}

}
\author{
\strong{Maintainer}: Emilia Jarochowska \email{e.jarochowska@uni-muenster.de} (\href{https://orcid.org/0000-0001-8937-9405}{ORCID})

Authors:
\itemize{
  \item Niklas Hohmann \email{N.H.Hohmann@uu.nl} (\href{https://orcid.org/0000-0003-1559-1838}{ORCID})
}

}
\keyword{internal}
