% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratPal-package.R
\docType{package}
\name{StratPal-package}
\alias{StratPal}
\alias{StratPal-package}
\title{StratPal: Stratigraphic Paleobiology Modeling Pipelines}
\description{
Models for stratigraphic paleobiology, including niche modeling and age-depth modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mindthegap-erc.github.io/StratPal/ }
  \item \url{https://github.com/MindTheGap-ERC/StratPal}
  \item Report bugs at \url{https://github.com/MindTheGap-ERC/StratPal/issues}
}

}
\author{
\strong{Maintainer}: Niklas Hohmann \email{N.H.Hohmann@uu.nl} (\href{https://orcid.org/0000-0003-1559-1838}{ORCID})

}
\keyword{internal}
